test: put every test file under the name of the source it covers
This stack had introduced eighteen test files whose basename had no source file
beside it, which is the convention PR #269 cleared the repository of.

It is not a style rule. The cost that established it was concrete: a file named
for a theme rather than a module matched a gitignore pattern meant for coverage
artifacts, and six hundred lines of real tests sat untracked for months while CI
never ran them. A name that follows the source cannot be swallowed that way, and
a reader can go from either direction without searching.

Each orphan went to the file matching the source it exercises, by what it
actually tests rather than by what its name suggested: cardinality_test.go
covered the metric attribute list in middleware.go, not the View in
internal/telemetry that shares the word.

One merge exposed a real constraint rather than a mechanical one.
logging_correlation_test.go was in the external toolutil_test package and
logging_test.go is internal, so combining them imported the package into its own
test. It is internal now, which it should have been: it exercises an exported
function of the package it lives in and needed no distance from it.

Everything was re-run with -shuffle=on, which is the check this kind of move
calls for: relocating tests is how order dependencies between them surface, and
ordinary go test hides that whole class.
